Restauracion-Imagenes/requirements.txt
streamlit
diffusers
transformers
torch
opencv-python
Pillow
open_clip_torch
requests
pyyaml
